requirements.txt
Django==5.1.3
channels==4.1.0
mysqlclient==2.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0